test_bk_resource/requirements.txt
amqp==5.2.0
appnope==0.1.4
arrow==1.3.0
asgiref==3.8.1
aspy.refactor-imports==3.0.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
binaryornot==0.4.4
bk-iam==1.1.19
bk_resource==0.4.12
bkapi-client-core==1.2.0
bkapi-component-open==2.1.0
bkoauth==0.1.0
blueapps==4.13.2
cachetools==3.1.1
celery==5.2.3
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cookiecutter==2.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cruft==2.15.0
cryptography==43.0.0
curlify==2.2.1
decorator==5.1.1
distlib==0.3.8
Django==3.2.25
django-celery-beat==2.2.0
django-celery-results==2.4.0
django-cors-headers==3.7.0
django-cors-middleware==1.3.1
django-environ==0.10.0
django-filter==2.4.0
django-jsonfield-backport==1.0.3
django-redis==5.2.0
django-rest-framework-condition==0.1.1
django-timezone-field==4.2.3
djangorestframework==3.12.4
drf-spectacular==0.12.0
drf-yasg==1.20.0
entrypoints==0.3
environ==1.0
execnet==2.1.1
executing==2.0.1
filelock==3.15.4
flake8==3.7.8
gitdb==4.0.11
GitPython==3.1.43
gunicorn==22.0.0
identify==2.6.0
idna==3.7
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.0.0
ipython==8.10.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
JSON-log-formatter==0.4.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.6.1
mdurl==0.1.2
mysqlclient==2.1.1
nodeenv==1.9.1
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==2.9.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pycodestyle==2.5.0
pycparser==2.22
pycryptodome==3.20.0
pyflakes==2.1.1
Pygments==2.18.0
pyinstrument==3.4.2
pyinstrument_cext==0.2.4
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==2.2.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.4.0
pytest-forked==1.6.0
pytest-html==3.1.1
pytest-metadata==2.0.4
pytest-xdist==2.4.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
raven==6.10.0
redis==5.0.0
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
seed-isort-config==1.9.2
sentry-sdk==1.28.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sqlparse==0.5.1
stack-data==0.6.3
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
traitlets==5.14.3
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
Werkzeug==2.3.8
whitenoise==5.2.0
xmlrunner==1.7.7
zipp==3.19.2
